+++ packages/liveblocks-react-lexical/src/liveblocks-plugin-provider.tsx
@@ -64,6 +64,8 @@ function useYjsProvider() {
  * - `loading`: Once the editor state has been requested by LiveblocksPlugin.
  * - `synchronizing`: Not working yet! Will be available in a future release.
  * - `synchronized`:  The editor state is sync with Liveblocks servers.
+ *
+ * @deprecated Prefer `useIsEditorReady` or `useSyncStatus` (from @liveblocks/react)
  */
 export function useEditorStatus(): EditorStatus {
   const provider = useYjsProvider();
@@ -85,6 +87,32 @@ export function useEditorStatus(): EditorStatus {
     }
     return provider.getStatus();
   }, [provider]);
+
+  return useSyncExternalStore(subscribe, getSnapshot, getSnapshot);
+}
+
+/**
+ * Returns whether the editor has loaded the initial text contents from the
+ * server and is ready to be used.
+ */
+export function useIsEditorReady(): boolean {
+  const yjsProvider = useYjsProvider();
+
+  const getSnapshot = useCallback(() => {
+    const status = yjsProvider?.getStatus();
+    return status === "synchronizing" || status === "synchronized";
+  }, [yjsProvider]);
+
+  const subscribe = useCallback(
+    (callback: () => void) => {
+      if (yjsProvider === undefined) return () => {};
+      yjsProvider.on("status", callback);
+      return () => {
+        yjsProvider.off("status", callback);
+      };
+    },
+    [yjsProvider]
+  );
 
   return useSyncExternalStore(subscribe, getSnapshot, getSnapshot);
 }